requirements.txt
chromadb
rank_bm25
networkx
pyyaml
requests
openai
pydantic
